feat ✨: surface Community Credit entries in the journal
Credit postings get their own `Credit` category, badge colour and filter pill,
plus readable activity labels — the two legs of one repayment installment read
differently, keyed off the interest leg's `Interest Expense` debit. The funded
sweep stays a `Transfer`: it is a Credit → Bank internal move, not a credit
event.

The category vocabulary moves out of `ledgerPresenter` into its own
`ledgerCategory` module, the same split `ledgerCurrency` already got, so the
presenter stays focused on turning entries into table rows. The presenter
re-exports it, so callers are unaffected.
