+++ app/src/utils/accounting/ledgerPresenter.ts
@@ -11,63 +11,31 @@ import { mergeBankFees } from './mergeBankFees'
 import { flattenLedgerRows } from './payrollGrouping'
 import { filterLedgerByCurrency } from './ledgerCurrency'
 import { formatAmountWithPrecision } from '@/utils/currencyUtil'
-import type { LedgerEntry, UseCase } from './ledgerEntry'
+import {
+  badgeClassOf,
+  categoryOf,
+  FEE_ACCOUNT,
+  FEE_FILTER,
+  type LedgerCategory
+} from './ledgerCategory'
+import type { LedgerEntry } from './ledgerEntry'
 import type { TokenId } from '@/constant'
 
 // Currency derivation / filtering lives in its own module, re-exported here.
 export { entryCurrency, ledgerCurrencies, filterLedgerByCurrency } from './ledgerCurrency'
+// So do the category vocabulary and its badges — see ./ledgerCategory.
+export {
+  badgeClassOf,
+  categoryOf,
+  CATEGORY_BADGE,
+  FEE_ACCOUNT,
+  FEE_FILTER,
+  ledgerCategories,
+  type LedgerCategory
+} from './ledgerCategory'
 
 /** The empty activity carried by a posting's continuation (credit) and total rows. */
 const NO_ACTIVITY: ActivityCell = { kind: 'plain', text: '' }
-
-/** Exact chart-of-accounts label for a protocol-fee leg; drives badge + filter. */
-export const FEE_ACCOUNT = 'Transaction Fee Expense'
-
-export type LedgerCategory =
-  | 'Investment'
-  | 'Revenue'
-  | 'Trading'
-  | 'Transfer'
-  | 'Payroll'
-  | 'Expense'
-  | 'Dividend'
-  | 'Memo'
-
-/**
- * Soft badge classes per ledger category — one distinct theme colour each, so
- * the "Action" column reads at a glance (static strings so Tailwind keeps them).
- * Colours come from the project palette (see `assets/main.css`).
- */
-export const CATEGORY_BADGE: Record<LedgerCategory, string> = {
-  Investment: 'bg-secondary/10 text-secondary', // capital in — blue
-  Revenue: 'bg-success/10 text-success', // income earned — green
-  Trading: 'bg-info/10 text-info', // market activity — cyan
-  Transfer: 'bg-neutral/10 text-neutral', // internal move — neutral
-  Payroll: 'bg-warning/10 text-warning', // wage accrued / owed — amber
-  Expense: 'bg-error/10 text-error', // cost out — red
-  Dividend: 'bg-primary/10 text-primary', // profit distribution — green
-  Memo: 'bg-muted text-dimmed' // share-count note — grey
-}
-
-/**
- * The pseudo-category the Fee pill filters on — not a {@link LedgerCategory} (a
- * fee is a leg of a Transfer/Expense entry), so it's handled specially by
- * {@link filterLedgerEntries} / {@link presentLedger} rather than via `categoryOf`.
- */
-export const FEE_FILTER = 'Fee'
-
-/** Ledger filter categories shown as pills (in design order). */
-export const ledgerCategories: Array<LedgerCategory | 'All' | typeof FEE_FILTER> = [
-  'All',
-  'Investment',
-  'Revenue',
-  'Trading',
-  'Transfer',
-  'Payroll',
-  'Expense',
-  'Dividend',
-  FEE_FILTER
-]
 
 /** The toggleable ledger table columns (keys match the table's cell slots). */
 export type LedgerColumnKey =
@@ -162,41 +130,6 @@ export interface LedgerView {
   rows: LedgerRow[]
   total: string
   entryCount: number
-}
-
-/**
- * Badge classes for a ledger entry's "Action" pill. Normally one colour per
- * category, but the two payroll use cases are split so the journal shows at a
- * glance whether a wage was merely **accrued** (submitted, still owed — amber)
- * or **settled** (withdrawn, actually paid out — green).
- */
-export function badgeClassOf(entry: LedgerEntry): string {
-  // A settled wage (UC-CASH-03 — withdrawn / actually paid out) reads as cyan,
-  // distinct from a wage merely accrued (UC-CASH-02 — submitted, still owed),
-  // which keeps the category's amber. Every other entry takes its category colour.
-  if (entry.useCase === 'UC-CASH-03') return 'bg-accent/10 text-accent'
-  return CATEGORY_BADGE[categoryOf(entry)]
-}
-
-/** The display category a ledger entry falls under, from its use case. */
-export function categoryOf(entry: LedgerEntry): LedgerCategory {
-  const byUseCase: Partial<Record<UseCase, LedgerCategory>> = {
-    'UC-BANK-01': 'Investment',
-    'UC-SDR-01': 'Investment',
-    'UC-MEMBER-01': 'Investment',
-    'UC-BANK-02': 'Revenue',
-    'CASH-IN': 'Revenue',
-    'UC-CASH-02': 'Payroll',
-    'UC-CASH-03': 'Payroll',
-    'UC-EXP-01': 'Expense',
-    'CASH-OUT': 'Expense',
-    'UC-INV-01': 'Dividend',
-    'DEFAULT-D': 'Investment',
-    FEE: 'Expense',
-    INTERNAL: 'Transfer',
-    'UC-BANK-03': 'Transfer'
-  }
-  return byUseCase[entry.useCase] ?? 'Transfer'
 }
 
 /** The Devise / Quantité / Taux columns of one token move (spec §2). */
